pyproject.toml
[tool.poetry]
name = "p21-fatsapi"
version = "0.1.0"
description = ""
authors = ["nmunisa06 <munisa06nm@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
jinja2 = "^3.1.4"
sqlmodel = "^0.0.22"
sqladmin = "^0.19.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"
fastapi-storages = "^0.3.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
pillow = "^10.4.0"
sqlalchemy-file = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
